Adding in a components directory and moving tags over.
